% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_metrics.R
\name{m_index}
\alias{m_index}
\title{M index}
\usage{
m_index(x)
}
\arguments{
\item{x}{SpatRaster or matrix containing two values: -1 and 1}
}
\value{
A numeric vector
}
\description{
Calculates m index -- a sum of cell’s values
over the entire site divided by the number of cell in the site.
m has a range from -1 (site completely dominated by the -1 values) to
1 (site completely dominated by the 1 values).
}
\examples{
data(r_start, package = "spatialising")
m_index(r_start)
ts1 = kinetic_ising(r_start, B = -0.3, J = 0.7)
m_index(ts1)
ts10 = kinetic_ising(r_start, B = -0.3, J = 0.7, updates = 2)
m_index(ts10)

# library(terra)
# r1 = rast(system.file("raster/r_start.tif", package = "spatialising"))
# m_index(r1)
# r2 = kinetic_ising(r1, B = -0.3, J = 0.7)
# m_index(r2)
}
