% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_metrics.R
\name{magnetization}
\alias{magnetization}
\title{Magnetization}
\usage{
magnetization(x)
}
\arguments{
\item{x}{SpatRaster or matrix containing two values: -1 and 1}
}
\value{
A numeric vector
}
\description{
Calculated magnetization
}
\examples{
data(r_start, package = "spatialising")
magnetization(r_start)
ts1 = spatial_ising(r_start, B = -0.3, J = 0.7)
magnetization(ts1)
ts10 = spatial_ising(r_start, B = -0.3, J = 0.7, updates = 2)
magnetization(ts10)

# library(terra)
# r1 = rast(system.file("raster/r_start.tif", package = "spatialising"))
# magnetization(r1)
# r2 = spatial_ising(r1, B = -0.3, J = 0.7)
# magnetization(r2)
}
