% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_end}
\alias{r_end}
\title{An example matrix object}
\format{
A matrix
}
\usage{
data(r_end)
}
\description{
A matrix has 50 columns and 50 rows. The matrix contains two values: -1 and 1.
This object was created with \code{r_end = kinetic_ising(r_start, B = -0.3, J = 0.7)}
}
\keyword{datasets}
