% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r_start}
\alias{r_start}
\title{An example matrix object}
\format{
A matrix
}
\usage{
data(r_start)
}
\description{
A matrix has 50 columns and 50 rows. The matrix contains two values: -1 and 1.
}
\keyword{datasets}
