% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{r_start.tif}
\alias{r_start.tif}
\title{An example binary raster}
\format{
A raster file
}
\description{
A raster file covering an area of 50x50 cells. The raster file contains two values: -1 and 1.
\code{system.file("raster/r_start.tif", package = "spatialising")}
}
