% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ising.R
\name{spatial_ising}
\alias{spatial_ising}
\title{Ising model for spatial data}
\usage{
spatial_ising(
  x,
  B,
  J,
  updates = 1,
  iter,
  rule = "metropolis",
  inertia = 0,
  version = 1,
  progress = TRUE
)
}
\arguments{
\item{x}{SpatRaster or matrix containing two values: -1 and 1}

\item{B}{External pressure}

\item{J}{Peer pressure - it regulates a degree of local harmonization}

\item{updates}{Specifies how many sets of iterations are performed on the input object.
The output of this function has as many layers as the \code{updates} value.}

\item{iter}{Specifies how many iterations are performed on the input object.
By default it equals to the number of values in the input object.}

\item{rule}{description}

\item{inertia}{description}

\item{version}{By default, \code{1}, the \code{x} object is converted into a matrix
(fast, but can be memory consuming); \code{version = 2} has a lower RAM impact, but
is much slower}

\item{progress}{TRUE/FALSE}
}
\value{
Object of the same class as \code{x} with the number of layers specified by \code{updates}
}
\description{
Performs simulations based on the given parameters of the Ising model
}
\examples{
data(r_start, package = "spatialising")
ts1 = spatial_ising(r_start, B = -0.3, J = 0.7)
ts10 = spatial_ising(r_start, B = -0.3, J = 0.7, updates = 10)

library(terra)
r1 = rast(system.file("raster/r_start.tif", package = "spatialising"))
plot(r1)
r2 = spatial_ising(r1, B = -0.3, J = 0.7)
plot(r2)

# ri1 = spatial_ising(r1, B = -0.3, J = 0.7, updates = 9)
# plot(ri1)

# ri2 = spatial_ising(r1, B = 0.3, J = 0.7, updates = 9)
# plot(ri2)

# ri3 = spatial_ising(r1, B = -0.3, J = 0.4, updates = 9)
# plot(ri3)
}
\references{
Ising, E., 1924. Beitrag zur theorie des ferro-und paramagnetismus. Ph.D. thesis, Grefe & Tiedemann.

Onsager, L., 1944. Crystal statistics. I. A two-dimensional model with an order-disorder transition. Physical Review 65 (3-4), 117.

Brush, S. G., 1967. History of the Lenz-Ising model. Reviews of modern physics 39 (4), 883.

Cipra, B. A., 1987. An introduction to the Ising model. The American Mathematical Monthly 94 (10), 937–959.
}
