% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_ising_ensemble.R
\name{spatial_ising_ensemble}
\alias{spatial_ising_ensemble}
\title{Ensemble of Ising models for spatial data}
\usage{
spatial_ising_ensemble(runs, ...)
}
\arguments{
\item{runs}{A number of simulations to perform}

\item{...}{Arguments for \code{\link[=spatial_ising]{spatial_ising()}}}
}
\value{
A list of objects of the same class as \code{x}
}
\description{
Creates an ensemble of simulations based on the given parameters of the Ising model
}
\examples{
data(r_start, package = "spatialising")
l = spatial_ising_ensemble(100, r_start, B = -0.3, J = 0.7)

# library(terra)
# r1 = rast(system.file("raster/r_start.tif", package = "spatialising"))
# plot(r1)
# r2 = spatial_ising_ensemble(100, r1, B = -0.3, J = 0.7)
}
