% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialising-package.R
\docType{package}
\name{spatialising-package}
\alias{spatialising}
\alias{spatialising-package}
\title{spatialising: Ising Model for Spatial Data}
\description{
Performs simulations of binary spatial raster data using the Ising model (Ising (1925) \doi{10.1007/BF02980577}; Onsager (1944) \doi{10.1103/PhysRev.65.117}). It allows to set a few parameters that represent internal and external pressures, and the number of simulations (Stepinski and Nowosad (2023) \doi{10.1098/rsos.231005}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jakubnowosad.com/spatialising/}
  \item Report bugs at \url{https://github.com/Nowosad/spatialising/issues}
}

}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})

}
\keyword{internal}
